Eric branch

+++ src/main/java/com/shield/daos/user/UserDAOImp.java
@@ -143,7 +143,7 @@ public class UserDAOImp implements UserDAO {
                         resultSet.getInt("claim_id"),
                         resultSet.getInt("user_id"),
                         resultSet.getInt("employee_id"),
-                        resultSet.getInt("amount"),
+                        resultSet.getFloat("amount"),
                         resultSet.getString("description"),
                         resultSet.getDate("date_of_occurrence").toString(),
                         resultSet.getString("location_of_occurrence"),
